+++ README.md
@@ -32,6 +32,6 @@ app 모듈에 있는 모든 컨트롤러를 어노테이션 기반 MVC로 변경
 ## 체크리스트
 
 - [x] 힌트에서 제공한 UserController 컨트롤러가 json 형태로 응답을 반환한다.
-- [ ] 레거시 코드를 삭제하고 서버를 띄워도 정상 동작한다.
+- [x] 레거시 코드를 삭제하고 서버를 띄워도 정상 동작한다.
 
 ## 피드백
