--- app/src/main/java/com/techcourse/v2/LoginController.java
@@ -1,57 +1,0 @@
-package com.techcourse.v2;
-
-import com.techcourse.domain.User;
-import com.techcourse.repository.InMemoryUserRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.support.RequestMethod;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@Controller
-public class LoginController {
-
-    private static final Logger log = LoggerFactory.getLogger(LoginController.class);
-
-    @RequestMapping(value = "/login", method = RequestMethod.POST)
-    public ModelAndView login(HttpServletRequest request, HttpServletResponse response) {
-        if (UserSession.isLoggedIn(request.getSession())) {
-            return new ModelAndView(new JspView("redirect:/index.jsp"));
-        }
-
-        String path = InMemoryUserRepository.findByAccount(request.getParameter("account"))
-                .map(user -> {
-                    log.info("User : {}", user);
-                    return loginUser(request, user);
-                })
-                .orElse("redirect:/401.jsp");
-
-        return new ModelAndView(new JspView(path));
-    }
-
-    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
-    public ModelAndView showLoginView(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        String path = UserSession.getUserFrom(req.getSession())
-                .map(user -> {
-                    log.info("logged in {}", user.getAccount());
-                    return "redirect:/index.jsp";
-                })
-                .orElse("/login.jsp");
-
-        return new ModelAndView(new JspView(path));
-    }
-
-    private String loginUser(final HttpServletRequest request, final User user) {
-        if (user.checkPassword(request.getParameter("password"))) {
-            final var session = request.getSession();
-            session.setAttribute(UserSession.SESSION_KEY, user);
-            return "redirect:/index.jsp";
-        } else {
-            return "redirect:/401.jsp";
-        }
-    }
-}
