+++ mvc/src/test/java/nextstep/mvc/HandlerAdapterRegistryTest.java
@@ -7,11 +7,9 @@ import static org.mockito.Mockito.mock;
 import java.util.NoSuchElementException;
 import nextstep.mvc.controller.tobe.HandlerExecution;
 import nextstep.mvc.controller.tobe.adapter.AnnotationHandlerAdapter;
-import nextstep.mvc.controller.tobe.adapter.ManualHandlerAdapter;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import samples.TestManualController;
 
 class HandlerAdapterRegistryTest {
 
@@ -20,7 +18,6 @@ class HandlerAdapterRegistryTest {
     @BeforeEach
     void setup() {
         handlerAdapterRegistry.addHandlerAdapter(new AnnotationHandlerAdapter());
-        handlerAdapterRegistry.addHandlerAdapter(new ManualHandlerAdapter());
     }
 
     @DisplayName("어노테이션 핸들러 어댑터를 찾을 수 있다.")
@@ -34,19 +31,6 @@ class HandlerAdapterRegistryTest {
 
         //then
         assertThat(adapter.getClass()).isEqualTo(AnnotationHandlerAdapter.class);
-    }
-
-    @DisplayName("메뉴얼 핸들러 어댑터를 찾을 수 있다.")
-    @Test
-    void findAdapter_Manual() {
-        //given
-        TestManualController handler = new TestManualController("/");
-
-        //when
-        HandlerAdapter adapter = handlerAdapterRegistry.findAdapter(handler);
-
-        //then
-        assertThat(adapter.getClass()).isEqualTo(ManualHandlerAdapter.class);
     }
 
     @DisplayName("일치하는 핸들러가 없을 시 예외가 발생한다.")
